requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
anyio==4.11.0
async-timeout==5.0.1
attrs==25.4.0
beautifulsoup4==4.14.2
blessed==1.23.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.0
cmake==3.25.0
contourpy==1.3.2
cycler==0.12.1
einops==0.8.1
filelock==3.19.1
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.10.0
gdown==5.2.0
geotorch==0.3.0
gpustat==1.1.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
kiwisolver==1.4.9
lit==15.0.7
loguru==0.7.3
lor-deckcodes==5.0.0
MarkupSafe==2.1.5
matplotlib==3.10.7
mpmath==1.3.0
multidict==6.7.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-ml-py==13.580.82
nvidia-nccl-cu11==2.19.3
nvidia-nvtx-cu11==11.8.86
opencv-python==4.12.0.88
opt_einsum==3.4.0
pandas==2.3.3
pillow==12.0.0
POT==0.9.6.post1
prefetch_generator==1.0.3
propcache==0.4.1
wilds==2.0.0
pybind11==3.0.1
pyot==6.0.9
pyparsing==3.2.5
PySocks==1.7.1
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.6.2
scipy==1.15.3
setproctitle==1.3.2
shellingham==1.5.4
sniffio==1.3.1
soupsieve==2.8
sympy==1.14.0
timm==1.0.22
tokenizers==0.22.1
torch==2.2.0+cu118
torchaudio==2.2.0+cu118
torchdiffeq==0.2.5
torchvision==0.17.0+cu118
tqdm==4.67.1
triton==2.2.0
typer-slim==0.20.0
tzdata==2025.2
urllib3==2.5.0
yacs==0.1.8
yarl==1.22.0
